requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
aiohttp==3.9.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
